backend/requirement.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
fastapi==0.114.0
idna==3.8
pillow==10.4.0
pydantic==2.9.0
pydantic_core==2.23.2
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
